feat(Core/Maps): Improvements to Cinematic function  (#10765)
